azure_fn/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
sentry-sdk==1.4.3
requests==2.26.0
psycopg2-binary==2.9.2
